requirements.txt
numpy
spatialmath-python
pyrender
efficientnet_pytorch
albumentations
matplotlib


